+++ plugins/react/frontend/components/common/Button/index.js
@@ -1,0 +1,14 @@
+/**
+ * Button
+ *
+ * Render a button
+ */
+
+import React from 'react';
+import styles from './styles.css';
+
+const Button = (props) => (
+  <button {...props} className={`${styles.base} ${props.className}`} />
+);
+
+export default Button;